Rename `ImageColorIcon` to `ColorOverlayIcon` to reflect it's no longer based on an `ImageIcon`
